+++ pubspec.yaml
@@ -37,6 +37,10 @@ dependencies:
 
   flutter_config: ^2.0.0
   flutter_inappwebview: ^5.3.2
+  flutter_slidable: ^0.6.0
+  timeago: ^3.1.0
+  material_floating_search_bar: ^0.3.6
+  google_maps_flutter: ^2.1.0
   fluttertoast: ^8.0.8
   geocoding: ^2.0.1
   google_maps_flutter: ^2.1.0
